Added CRC code, 'locations', lzss
The CRC routines have been added to the main code, and replace the previous
random routines.

A 'locations' word has been added to the metacompiler, which prints out all
words defined in the target dictionary by the metacompiler, as well as their
locations.

'lzss.c' has been added, which is a potential way of decreasing the image size
even more - this would need to be translated to Forth first.

A word for performing a built in self test has been added, using the CRC and
length field to perform some basic checks. It is comment out for the moment.

There have been some other minor changes.

+++ forth.c
@@ -159,7 +159,7 @@ int forth(forth_t *h, FILE *in, FILE *out, const char *block)
 	}
 finished:
 	h->pc = pc; h->sp = sp; h->rp = rp; h->t = t;
-	return t;
+	return (int16_t)t;
 }
 
 int main(int argc, char **argv)
